fix: address operator user credit ledger review
