+++ src/cook-web/src/app/admin/operations/users/[user_bid]/page.tsx
@@ -6,25 +6,14 @@ import { useParams, useRouter } from 'next/navigation';
 import { CircleHelp } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
 import api from '@/api';
-import AdminDateRangeFilter from '@/app/admin/components/AdminDateRangeFilter';
-import { AdminPagination } from '@/app/admin/components/AdminPagination';
 import AdminTooltipText from '@/app/admin/components/AdminTooltipText';
 import { formatAdminCredits } from '@/app/admin/lib/numberFormat';
-import { ClearableTextInput } from '@/app/admin/operations/orders/orderUiShared';
 import { useEnvStore } from '@/c-store';
 import type { EnvStoreState } from '@/c-types/store';
 import ErrorDisplay from '@/components/ErrorDisplay';
 import Loading from '@/components/loading';
 import { Button } from '@/components/ui/Button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/Card';
-import { Label } from '@/components/ui/Label';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/Select';
 import {
   Table,
   TableBody,
@@ -50,14 +39,17 @@ import { normalizeLoginMethodLabelKey } from '../loginMethodUtils';
 import type {
   AdminOperationUserCourseItem,
   AdminOperationUserCreditFilters,
-  AdminOperationUserCreditGrantSourceFilter,
   AdminOperationUserCreditSummary,
   AdminOperationUserCreditsResponse,
   AdminOperationUserDetailResponse,
-  AdminOperationUserCreditTypeFilter,
-  AdminOperationUserCreditUsageModeFilter,
 } from '../../operation-user-types';
 import useOperatorGuard from '../../useOperatorGuard';
+import UserCreditLedgerTab from './UserCreditLedgerTab';
+import {
+  createUserCreditFilters,
+  FILTER_ALL_OPTION,
+  sanitizeCreditFiltersByType,
+} from './creditFilterUtils';
 
 type ErrorState = { message: string; code?: number };
 type DetailTab = 'credits' | 'learning' | 'created';
@@ -65,7 +57,6 @@ type DetailTab = 'credits' | 'learning' | 'created';
 const CREDITS_PAGE_SIZE = 10;
 const EMPTY_VALUE = '--';
 const DEFAULT_VISIBLE_COURSE_COUNT = 10;
-const FILTER_ALL_OPTION = 'all';
 const DEFAULT_CREDIT_SUMMARY: AdminOperationUserCreditSummary = {
   available_credits: '',
   subscription_credits: '',
@@ -81,18 +72,6 @@ const createEmptyCreditsResponse = (): AdminOperationUserCreditsResponse => ({
   page_size: CREDITS_PAGE_SIZE,
   total: 0,
 });
-const createUserCreditFilters = (): AdminOperationUserCreditFilters => ({
-  creditType: FILTER_ALL_OPTION,
-  grantSource: FILTER_ALL_OPTION,
-  courseQuery: '',
-  usageMode: FILTER_ALL_OPTION,
-  startTime: '',
-  endTime: '',
-});
-type OperatorUsersTranslator = (
-  key: string,
-  options?: { defaultValue?: string },
-) => string;
 const EMPTY_DETAIL: AdminOperationUserDetailResponse = {
   user_bid: '',
   mobile: '',
@@ -261,29 +240,6 @@ const formatLearningProgress = (
   return `${progressPercent}% (${completedLessonCount}/${totalLessonCount})`;
 };
 
-const resolveCreditLedgerLabel = (
-  tOperationsUsers: OperatorUsersTranslator,
-  type: 'creditLedgerTypeLabels' | 'creditLedgerSourceLabels',
-  displayCode: string,
-  fallbackCode: string,
-): string => {
-  const normalizedDisplayCode = displayCode.trim();
-  if (normalizedDisplayCode) {
-    return tOperationsUsers(`detail.${type}.${normalizedDisplayCode}`, {
-      defaultValue: normalizedDisplayCode,
-    });
-  }
-  return fallbackCode.trim() || EMPTY_VALUE;
-};
-
-const resolveCreditLedgerNote = (note: string): string => {
-  const normalizedNote = note.trim();
-  if (normalizedNote) {
-    return normalizedNote;
-  }
-  return EMPTY_VALUE;
-};
-
 const CourseTable = ({
   title,
   courses,
@@ -411,478 +367,6 @@ const CourseTable = ({
               {expanded ? t('common.core.collapse') : t('common.core.expand')}
             </Button>
           </div>
-        ) : null}
-      </CardContent>
-    </Card>
-  );
-};
-
-const sanitizeCreditFiltersByType = (
-  filters: AdminOperationUserCreditFilters,
-): AdminOperationUserCreditFilters => {
-  if (filters.creditType === 'grant') {
-    return {
-      ...filters,
-      courseQuery: '',
-      usageMode: FILTER_ALL_OPTION,
-    };
-  }
-  if (filters.creditType === 'consume') {
-    return {
-      ...filters,
-      grantSource: FILTER_ALL_OPTION,
-    };
-  }
-  if (filters.creditType === 'other') {
-    return {
-      ...filters,
-      grantSource: FILTER_ALL_OPTION,
-      courseQuery: '',
-      usageMode: FILTER_ALL_OPTION,
-    };
-  }
-  return createUserCreditFilters();
-};
-
-const CreditLedgerFilters = ({
-  filtersDraft,
-  loading,
-  onChange,
-  onSearch,
-  onReset,
-}: {
-  filtersDraft: AdminOperationUserCreditFilters;
-  loading: boolean;
-  onChange: (filters: AdminOperationUserCreditFilters) => void;
-  onSearch: () => void;
-  onReset: () => void;
-}) => {
-  const { t } = useTranslation();
-  const { t: tOperationsUsers } = useTranslation('module.operationsUser');
-  const showGrantFilters = filtersDraft.creditType === 'grant';
-  const showConsumeFilters = filtersDraft.creditType === 'consume';
-  const showOtherFilters = filtersDraft.creditType === 'other';
-
-  return (
-    <form
-      className='rounded-xl border border-border bg-muted/20 p-3'
-      onSubmit={event => {
-        event.preventDefault();
-        onSearch();
-      }}
-    >
-      <div className='flex flex-col gap-3 xl:flex-row xl:items-end'>
-        <div className='flex flex-col gap-2 xl:w-[240px] xl:flex-none'>
-          <Label className='text-xs font-medium text-muted-foreground'>
-            {tOperationsUsers('detail.creditLedgerFilters.type')}
-          </Label>
-          <Select
-            value={filtersDraft.creditType}
-            onValueChange={value =>
-              onChange(
-                sanitizeCreditFiltersByType({
-                  ...filtersDraft,
-                  creditType: value as AdminOperationUserCreditTypeFilter,
-                }),
-              )
-            }
-          >
-            <SelectTrigger className='h-9'>
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value={FILTER_ALL_OPTION}>
-                {tOperationsUsers('detail.creditLedgerFilters.typeOptions.all')}
-              </SelectItem>
-              <SelectItem value='consume'>
-                {tOperationsUsers(
-                  'detail.creditLedgerFilters.typeOptions.consume',
-                )}
-              </SelectItem>
-              <SelectItem value='grant'>
-                {tOperationsUsers(
-                  'detail.creditLedgerFilters.typeOptions.grant',
-                )}
-              </SelectItem>
-              <SelectItem value='other'>
-                {tOperationsUsers(
-                  'detail.creditLedgerFilters.typeOptions.other',
-                )}
-              </SelectItem>
-            </SelectContent>
-          </Select>
-        </div>
-
-        {showGrantFilters ? (
-          <div className='flex flex-1 flex-col gap-2'>
-            <Label className='text-xs font-medium text-muted-foreground'>
-              {tOperationsUsers('detail.creditLedgerFilters.grantSource')}
-            </Label>
-            <Select
-              value={filtersDraft.grantSource}
-              onValueChange={value =>
-                onChange({
-                  ...filtersDraft,
-                  grantSource:
-                    value as AdminOperationUserCreditGrantSourceFilter,
-                })
-              }
-            >
-              <SelectTrigger className='h-9'>
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value={FILTER_ALL_OPTION}>
-                  {tOperationsUsers(
-                    'detail.creditLedgerFilters.grantSourceOptions.all',
-                  )}
-                </SelectItem>
-                <SelectItem value='subscription'>
-                  {tOperationsUsers(
-                    'detail.creditLedgerFilters.grantSourceOptions.subscription',
-                  )}
-                </SelectItem>
-                <SelectItem value='topup'>
-                  {tOperationsUsers(
-                    'detail.creditLedgerFilters.grantSourceOptions.topup',
-                  )}
-                </SelectItem>
-                <SelectItem value='trial_subscription'>
-                  {tOperationsUsers(
-                    'detail.creditLedgerFilters.grantSourceOptions.trial_subscription',
-                  )}
-                </SelectItem>
-                <SelectItem value='manual'>
-                  {tOperationsUsers(
-                    'detail.creditLedgerFilters.grantSourceOptions.manual',
-                  )}
-                </SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-        ) : null}
-
-        {showConsumeFilters ? (
-          <div className='flex flex-1 flex-col gap-2'>
-            <Label className='text-xs font-medium text-muted-foreground'>
-              {tOperationsUsers('detail.creditLedgerFilters.course')}
-            </Label>
-            <ClearableTextInput
-              value={filtersDraft.courseQuery}
-              placeholder={tOperationsUsers(
-                'detail.creditLedgerFilters.coursePlaceholder',
-              )}
-              clearLabel={t('module.chat.lessonFeedbackClearInput')}
-              onChange={value =>
-                onChange({
-                  ...filtersDraft,
-                  courseQuery: value,
-                })
-              }
-            />
-          </div>
-        ) : null}
-
-        {showConsumeFilters ? (
-          <div className='flex flex-1 flex-col gap-2'>
-            <Label className='text-xs font-medium text-muted-foreground'>
-              {tOperationsUsers('detail.creditLedgerFilters.usageMode')}
-            </Label>
-            <Select
-              value={filtersDraft.usageMode}
-              onValueChange={value =>
-                onChange({
-                  ...filtersDraft,
-                  usageMode: value as AdminOperationUserCreditUsageModeFilter,
-                })
-              }
-            >
-              <SelectTrigger className='h-9'>
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value={FILTER_ALL_OPTION}>
-                  {tOperationsUsers(
-                    'detail.creditLedgerFilters.usageModeOptions.all',
-                  )}
-                </SelectItem>
-                <SelectItem value='learn'>
-                  {tOperationsUsers(
-                    'detail.creditLedgerFilters.usageModeOptions.learn',
-                  )}
-                </SelectItem>
-                <SelectItem value='listen'>
-                  {tOperationsUsers(
-                    'detail.creditLedgerFilters.usageModeOptions.listen',
-                  )}
-                </SelectItem>
-                <SelectItem value='ask'>
-                  {tOperationsUsers(
-                    'detail.creditLedgerFilters.usageModeOptions.ask',
-                  )}
-                </SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-        ) : null}
-
-        {showGrantFilters || showConsumeFilters || showOtherFilters ? (
-          <div
-            className={cn(
-              'flex flex-1 flex-col gap-2',
-              showOtherFilters && 'xl:w-[420px] xl:flex-none',
-            )}
-          >
-            <Label className='text-xs font-medium text-muted-foreground'>
-              {tOperationsUsers('detail.creditLedgerFilters.time')}
-            </Label>
-            <AdminDateRangeFilter
-              startValue={filtersDraft.startTime}
-              endValue={filtersDraft.endTime}
-              triggerAriaLabel={tOperationsUsers(
-                'detail.creditLedgerFilters.time',
-              )}
-              placeholder={tOperationsUsers(
-                'detail.creditLedgerFilters.timePlaceholder',
-              )}
-              resetLabel={t('module.order.filters.reset')}
-              clearLabel={t('module.chat.lessonFeedbackClearInput')}
-              onChange={({ start, end }) =>
-                onChange({
-                  ...filtersDraft,
-                  startTime: start,
-                  endTime: end,
-                })
-              }
-            />
-          </div>
-        ) : null}
-
-        <div className='flex min-h-9 shrink-0 items-center justify-start gap-2 xl:ml-auto xl:justify-end'>
-          <Button
-            type='button'
-            variant='outline'
-            className='h-9 px-4'
-            onClick={onReset}
-            disabled={loading}
-          >
-            {t('module.order.filters.reset')}
-          </Button>
-          <Button
-            type='submit'
-            className='h-9 px-4'
-            disabled={loading}
-          >
-            {t('module.order.filters.search')}
-          </Button>
-        </div>
-      </div>
-    </form>
-  );
-};
-
-const CreditLedgerTable = ({
-  filtersDraft,
-  loading,
-  error,
-  items,
-  pageIndex,
-  pageCount,
-  onFiltersChange,
-  onSearch,
-  onReset,
-  onPageChange,
-  onRetry,
-}: {
-  filtersDraft: AdminOperationUserCreditFilters;
-  loading: boolean;
-  error: ErrorState | null;
-  items: AdminOperationUserCreditsResponse['items'];
-  pageIndex: number;
-  pageCount: number;
-  onFiltersChange: (filters: AdminOperationUserCreditFilters) => void;
-  onSearch: () => void;
-  onReset: () => void;
-  onPageChange: (page: number) => void;
-  onRetry: () => void;
-}) => {
-  const { t, i18n } = useTranslation();
-  const { t: tOperationsUsers } = useTranslation('module.operationsUser');
-
-  if (error) {
-    return (
-      <div className='rounded-xl border border-border bg-white p-4 shadow-sm'>
-        <ErrorDisplay
-          errorCode={error.code || 0}
-          errorMessage={error.message}
-          onRetry={onRetry}
-        />
-      </div>
-    );
-  }
-
-  return (
-    <Card
-      className='shadow-sm'
-      data-testid='admin-operation-user-credit-ledger-card'
-    >
-      <CardContent className='space-y-4 pt-6'>
-        <CreditLedgerFilters
-          filtersDraft={filtersDraft}
-          loading={loading}
-          onChange={onFiltersChange}
-          onSearch={onSearch}
-          onReset={onReset}
-        />
-        <TooltipProvider delayDuration={150}>
-          <div
-            className='overflow-auto'
-            data-testid='admin-operation-user-credit-ledger-scroll'
-          >
-            <Table className='table-fixed'>
-              <colgroup>
-                <col className='w-[16%]' />
-                <col className='w-[13%]' />
-                <col className='w-[12%]' />
-                <col className='w-[10%]' />
-                <col className='w-[11%]' />
-                <col className='w-[14%]' />
-                <col className='w-[24%]' />
-              </colgroup>
-              <TableHeader>
-                <TableRow>
-                  <TableHead className='text-center'>
-                    {tOperationsUsers('detail.creditLedgerColumns.createdAt')}
-                  </TableHead>
-                  <TableHead className='text-center'>
-                    {tOperationsUsers('detail.creditLedgerColumns.entryType')}
-                  </TableHead>
-                  <TableHead className='text-center'>
-                    {tOperationsUsers('detail.creditLedgerColumns.sourceType')}
-                  </TableHead>
-                  <TableHead className='text-center'>
-                    {tOperationsUsers('detail.creditLedgerColumns.amount')}
-                  </TableHead>
-                  <TableHead className='text-center'>
-                    {tOperationsUsers(
-                      'detail.creditLedgerColumns.balanceAfter',
-                    )}
-                  </TableHead>
-                  <TableHead className='text-center'>
-                    {tOperationsUsers('detail.creditLedgerColumns.expiresAt')}
-                  </TableHead>
-                  <TableHead className='text-center'>
-                    {tOperationsUsers('detail.creditLedgerColumns.note')}
-                  </TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {loading ? (
-                  <TableEmpty colSpan={7}>
-                    {tOperationsUsers('detail.loadingCredits')}
-                  </TableEmpty>
-                ) : items.length ? (
-                  items.map(item => (
-                    <TableRow key={item.ledger_bid}>
-                      <TableCell className='max-w-0 overflow-hidden text-ellipsis whitespace-nowrap text-center'>
-                        <AdminTooltipText
-                          text={formatOperatorNaiveDateTime(item.created_at)}
-                          emptyValue={EMPTY_VALUE}
-                        />
-                      </TableCell>
-                      <TableCell className='max-w-0 overflow-hidden text-ellipsis whitespace-nowrap text-center'>
-                        <AdminTooltipText
-                          text={resolveCreditLedgerLabel(
-                            tOperationsUsers,
-                            'creditLedgerTypeLabels',
-                            item.display_entry_type,
-                            item.entry_type,
-                          )}
-                          emptyValue={EMPTY_VALUE}
-                        />
-                      </TableCell>
-                      <TableCell className='max-w-0 overflow-hidden text-ellipsis whitespace-nowrap text-center'>
-                        <AdminTooltipText
-                          text={resolveCreditLedgerLabel(
-                            tOperationsUsers,
-                            'creditLedgerSourceLabels',
-                            item.display_source_type,
-                            item.source_type,
-                          )}
-                          emptyValue={EMPTY_VALUE}
-                        />
-                      </TableCell>
-                      <TableCell className='max-w-0 overflow-hidden text-ellipsis whitespace-nowrap text-center'>
-                        <AdminTooltipText
-                          text={
-                            item.amount === '' ||
-                            item.amount === null ||
-                            item.amount === undefined
-                              ? ''
-                              : formatAdminCredits(
-                                  Number(item.amount),
-                                  i18n.language,
-                                )
-                          }
-                          emptyValue={EMPTY_VALUE}
-                        />
-                      </TableCell>
-                      <TableCell className='max-w-0 overflow-hidden text-ellipsis whitespace-nowrap text-center'>
-                        <AdminTooltipText
-                          text={
-                            item.balance_after === '' ||
-                            item.balance_after === null ||
-                            item.balance_after === undefined
-                              ? ''
-                              : formatAdminCredits(
-                                  Number(item.balance_after),
-                                  i18n.language,
-                                )
-                          }
-                          emptyValue={EMPTY_VALUE}
-                        />
-                      </TableCell>
-                      <TableCell className='max-w-0 overflow-hidden text-ellipsis whitespace-nowrap text-center'>
-                        <AdminTooltipText
-                          text={formatOperatorNaiveDateTime(item.expires_at)}
-                          emptyValue={EMPTY_VALUE}
-                        />
-                      </TableCell>
-                      <TableCell className='max-w-0 overflow-hidden text-ellipsis whitespace-nowrap text-center'>
-                        <AdminTooltipText
-                          text={resolveCreditLedgerNote(item.note)}
-                          emptyValue={EMPTY_VALUE}
-                        />
-                      </TableCell>
-                    </TableRow>
-                  ))
-                ) : (
-                  <TableEmpty colSpan={7}>
-                    {tOperationsUsers('detail.emptyCredits')}
-                  </TableEmpty>
-                )}
-              </TableBody>
-            </Table>
-          </div>
-        </TooltipProvider>
-
-        {pageCount > 1 ? (
-          <AdminPagination
-            pageIndex={pageIndex}
-            pageCount={pageCount}
-            onPageChange={onPageChange}
-            prevLabel={t('module.order.paginationPrev', 'Previous')}
-            nextLabel={t('module.order.paginationNext', 'Next')}
-            prevAriaLabel={t(
-              'module.order.paginationPrevAriaLabel',
-              'Go to previous page',
-            )}
-            nextAriaLabel={t(
-              'module.order.paginationNextAriaLabel',
-              'Go to next page',
-            )}
-            className='justify-end w-auto mx-0'
-          />
         ) : null}
       </CardContent>
     </Card>
@@ -1437,13 +921,14 @@ export default function AdminOperationUserDetailPage() {
                     value='credits'
                     className='mt-0'
                   >
-                    <CreditLedgerTable
+                    <UserCreditLedgerTab
                       filtersDraft={creditFiltersDraft}
                       loading={creditsLoading}
                       error={creditsError}
                       items={credits.items}
                       pageIndex={credits.page || creditsPageIndex}
                       pageCount={credits.page_count || 0}
+                      emptyValue={EMPTY_VALUE}
                       onFiltersChange={setCreditFiltersDraft}
                       onSearch={handleCreditSearch}
                       onReset={handleCreditReset}
